Split Android progress ViewModel tests
